Add second option for icon paths on Linux
This commit assesses whether the user has installed the icon packs into their /usr/share directory and, if not, it instead reverts to using the icons relative to the path of the currently executing script. This allows pyParrotZik to run in a standalone environment without relying on modifications to system files.